+++ test/perspective_test.exs
@@ -18,7 +18,7 @@ defmodule Image.Perspective.Test do
 
   test "warps an image to perspective" do
     image_path = image_path("warp_perspective.jpg")
-    validate_path = validate_path("warp_perspective.jpg")
+    validate_path = validate_path("warp/warp_perspective.jpg")
 
     {:ok, image} = Image.open(image_path)
     {:ok, result} = Image.warp_perspective(image, @warp_from, @warp_to)
@@ -31,7 +31,7 @@ defmodule Image.Perspective.Test do
 
   test "warps an image to a rectangular perspective" do
     image_path = image_path("warp_perspective.jpg")
-    validate_path = validate_path("warp_perspective_straighten.png")
+    validate_path = validate_path("warp/warp_perspective_straighten.png")
 
     {:ok, image} = Image.open(image_path)
     {:ok, destination, result} = Image.straighten_perspective(image, @warp_from)
@@ -45,7 +45,7 @@ defmodule Image.Perspective.Test do
 
   test "post-crop of a warped image" do
     image_path = image_path("warp_perspective.jpg")
-    validate_path = validate_path("warp_perspective_cropped.png")
+    validate_path = validate_path("warp/warp_perspective_cropped.png")
 
     {:ok, image} = Image.open(image_path)
     {:ok, result} = Image.warp_perspective(image, @warp_from, @warp_to)
@@ -54,5 +54,39 @@ defmodule Image.Perspective.Test do
 
     # Image.write! cropped, validate_path
     assert_images_equal(cropped, validate_path)
+  end
+
+  test "warp an image with an alpha band" do
+    image_path = image_path("image_with_alpha2.png")
+    validate_path = validate_path("warp/warped_image_with_alpha2.png")
+
+    {:ok, image} = Image.open(image_path)
+    {:ok, result} = Image.warp_perspective(image,
+      [{139, 125}, {826, 74}, {796, 559}, {155, 483}],
+      [{139, 125}, {815, 125}, {815, 528}, {139, 528}]
+    )
+
+    # Image.write! result, validate_path
+    {:ok, result} = Vix.Vips.Image.write_to_buffer(result, ".png")
+
+    assert_images_equal(result, validate_path)
+  end
+
+  test "warp an image with an alpha band making the added pixels transparent" do
+    image_path = image_path("image_with_alpha2.png")
+    validate_path = validate_path("warp/warped_image_with_alpha2_transparent.png")
+
+    {:ok, image} = Image.open(image_path)
+    {:ok, result} = Image.warp_perspective(image,
+      [{139, 125}, {826, 74}, {796, 559}, {155, 483}],
+      [{139, 125}, {815, 125}, {815, 528}, {139, 528}],
+      background: [1, 177, 64]
+    )
+    {:ok, result} = Image.chroma_key(result, color: [1, 177, 64], threshold: 0)
+
+    # Image.write! result, validate_path
+    {:ok, result} = Vix.Vips.Image.write_to_buffer(result, ".png")
+
+    assert_images_equal(result, validate_path)
   end
 end